+++ src/shared/commands/index.js
@@ -1,0 +1,4 @@
+import exec from './exec';
+import complete from './complete';
+
+export { exec, complete };
